faddil livrabre valide

+++ src/repositories/CompetenceRepository.ts
@@ -29,9 +29,50 @@ export class CompetenceRepository implements IRepository<Competence> {
     });
   }
 
-  async delete(id: number): Promise<void> {
-    await this.prisma.competence.delete({ where: { id } });
-  }
+//   async delete(id: number): Promise<void> {
+//     await this.prisma.competence.delete({ where: { id } });
+//   }
+
+    // async delete(id: number): Promise<void> {
+    
+    // await this.prisma.userCompetence.deleteMany({
+    //     where: { competenceId: id },
+    // });
+
+    
+    // await this.prisma.competence.delete({
+    //     where: { id },
+    // });
+    // }
+
+    // async delete(id: number): Promise<void> {
+    // const used = await this.prisma.userCompetence.count({
+    //     where: { competenceId: id },
+    // });
+
+    // if (used > 0) {
+    //     throw new Error("Impossible de supprimer : compétence encore utilisée par des utilisateurs.");
+    // }
+
+    // await this.prisma.competence.delete({ where: { id } });
+    // }
+
+        async delete(id: number): Promise<void> {
+            const usedInUser = await this.prisma.userCompetence.count({
+                 where: { competenceId: id },
+            });
+
+            const usedInRef = await this.prisma.refCompetence.count({
+                where: { competenceId: id },
+            });
+
+            if (usedInUser > 0 || usedInRef > 0) {
+             throw new Error("Impossible de supprimer : compétence encore utilisée.");
+            }
+
+             await this.prisma.competence.delete({ where: { id } });
+        }
+
 
   
   async findNiveauxByCompetence(id: number) {
